requirements.txt
flake8 == 6.*
google-api-python-client == 2.*
google-auth-httplib2 == 0.1.*
google-auth-oauthlib == 1.*
sib-api-v3-sdk == 7.*
